Get-Inspirational-Quote
Get-Inspirational-Quote pulls a random quote online using the quotes.rest api

IMPORTANT:
Unless the oauth authentication ability is added to the logic of this script, 
you will be limited to a few pulls every hour. This is an open query, used by many other people. 
Unless you are willing to sign up, don't abuse it.